requirements.txt
# Dependencias de Python
google-generativeai
yt-dlp
fpdf
python-dotenv
Pillow